requirements.txt
absl-py==2.0.0
altair==5.2.0
asttokens==2.4.1
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2022.6.15
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
flatbuffers==23.5.26
fonttools==4.44.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
idna==3.3
importlib-metadata==7.0.1
intel-openmp==2021.4.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
Js2Py==0.71
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.8
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
networkx==3.3
numpy==1.26.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==21.3
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
pipreqs==0.5.0
pipreqsnb==0.2.4
pipwin==0.5.2
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pycparser==2.21
pydeck==0.8.1b0
pygame==2.1.2
Pygments==2.17.2
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.28.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.13.0
seaborn==0.13.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
soupsieve==2.3.2.post1
SpeechRecognition @ file:///C:/Users/Mario/Downloads/SpeechRecognition-3.8.1-py2.py3-none-any.whl
stack-data==0.6.3
streamlit==1.31.1
sympy==1.12
tbb==2021.12.0
tenacity==8.2.3
thop==0.1.1.post2209072238
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==4.2
ultralytics==8.2.12
urllib3==1.26.9
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.17.0
